requirements.txt
# --- базовое окружение твоего бота ---
aiogram==3.12.0
httpx[http2]==0.27.0
python-dotenv==1.0.1
aiosqlite==0.20.0
cryptography==41.0.3
certifi==2024.8.30
pyperclip==1.9.0

# --- ускорители и оптимизации ---
uvloop>=0.22.0; sys_platform != "win32"
orjson>=3.10.0         # супербыстрый JSON
loguru>=0.7.2          # быстрая structured-логика без лишнего I/O
Pillow>=10.4.0         # если CPU x86_64 — замени на pillow-simd для +40% к скорости рендера
